Set `X-CURRENT-BLOCK-HASH` header on every request (#20)
Introduces a new optional argument `current-block-poll-interval`. When
that is set we create a `CurrentBlockStream` and use the current block
hash to populate the `X-CURRENT-BLOCK-HASH` header on every http
request.
I only made it optional to not have to introduce a bunch of mocking
logic in tests for a minor feature.

Also updates dependencies to the backend crates with the exception of
`solvers-dto`. Upgrading that requires a bunch of changes in unit tests
which are already covered in #19.