Handle zeroex legal error (#27)
We started receiving errors when trying to swap tokens that are
apparently, for some reason, banned from trading on 0x API.

Quick fix is to define a separate error variant for this type of error,
so it can be properly handled by our metrics system and not being
considered "unexpected".

The alternative would be to define a black list of tokens for 0x api for
which we know are banned and don't even send the requests for those. The
obvious problem with this is that we don't know the reason why these
tokens are banned and for how long. If they get unbanned we won't
notice.