Retry oneinch solver init (#36)
Recently the oneinch solver pod crashes during initialization more
frequently. This is due to rate limits during `OneInch::new()`.

This PR makes it so that we retry the initialization 10s because this is
usually just a temporary error.

While I was at it I unified the semantics of all `Solver::new()`
implementations to panic if a non recoverable error happens which makes
the code slightly nicer to read. (Only paraswap needed this change)