Classify 1Inch invalid token error correctly (#70)
We have seen some unactionable 1Inch solver errors over the past few
days. Looking at the logs, this is coming from API responses of the
form:

> { code: 400, description: "not valid token" }

indicating that the 1Inch API cannot handle certain tokens. We already
filter for insufficient liquidity (which also have a 400 code) so it
sees appropriate to label all 400 errors as `NotFound`