Fix 1Inch quoting issues (#81)
Currently the 1Inch solver returns a ton of 403 errors while quoting.
The reason is that the 1Inch API checks the `tx.origin` for legal
requirements and for quote requests we currently send the zero address.
Since that address obviously can't be the origin of any transaction they
refuse to return quotes for those.

To work around that we set `from` (our settlement contract) as the
`tx.origin` for quote requests. Since the calldata generated for that
will never get used to settle real orders this is okay.