Handle 403 errors for 1inch requests (#96)
From the original issue #95:

> The 1Inch API requires you to tell them which address wants to do the
trade. Due to legal requirements they do not serve some addresses. In
those cases the API call will return a 403 error.
Currently these get forwarded to the driver in a way that causes alerts
like this although they are to be expected and don't indicate an error
in the 1Inch API.

All the 403 errors are considered `NotFound` to reduce false alerts.

Fixes #95