Drop `useProtocolVersion` BalancerV3 request field (#120)
When switching to the Balancer SOR V3 GraphQL API, it was decided to
keep sending requests with the explicit protocol version 2 since it was
supposed to return better prices. And now an issue was observed where no
solutions can be found for liquidity pools v3. After trying to drop the
param or replace it with v3, the API returns valid solutions.

The field is optional, so let the Balancer protocol decide which version
to use.
<img width="500" alt="image"
src="https://github.com/user-attachments/assets/34ccfdf9-421e-4724-8454-6d4fe7f642cd"
/>

This probably requires a hotfix release.

Tested by running the solver locally.