Fixes for 0x v2 support (#125)
Contains 2 fixes that were implicit and ~not documented~ (not visible to
me) by 0x team in the migration document:

1. If sell/buy token is banned for legal reasons, the `422` error code
is returned instead of `451`. The code handling this error code is
added.
2. Apparently some quotes are returned without `sell_amount` in the
response. As confirmed by 0x team, this is not a bug, since
`liquidityAvailable` field should be `false` in that case. In any case,
I don't see how these responses could be valid, so I classified them as
`NotFound` errors.