Make balancer contracts non-mandatory (#133)
While deploying on Polygon, it turned out that Balancer is not fully
supported on this chain:
https://docs.balancer.fi/developer-reference/contracts/deployment-addresses/polygon.html#active-contracts

The `Contracts` struct, which contains protocol and Balancer contracts
in a single place, makes it impossible to spin up other solvers such as
0x, 1inch, in case Balancer is not supported on a particular chain, as
on Polygon.

This PR fixes it by fetching the Balancer contracts info only when its
solver is enabled.

Tested on Polygon.