Allow Balancer V2 usage separately from V3 (#134)
Only Balancer V2 is currently supported on Polygon. V2 and V3 are
coupled in the config, making it impossible to use the solver on this
chain. This PR fixes this by adding a config for enabled API versions.
BatchRouter won't be initialized if V3 is not specified in the config,
and VaultV2 won't work without the corresponding config. Absence of the
config means all versions are supported.

In case an API version is disabled and the solver starts sending errors
showing the disabled API version is now supported, it can be quickly
adjusted in the config.

Tested on Polygon: [cowprotocol/infrastructure@`af22e82`
(#2987)](https://github.com/cowprotocol/infrastructure/pull/2987/commits/af22e82b7935675588f86884283ad2e24ef88c62)