Minimum slippage (#140)
This PR was mainly implemented by Claude AI, so please review with
caution (I did a full review pass myself). The size of the PR is
reasonable if you discount the ~600 lines of added integration tests.

The goal of this PR is to address a noisy alert we are encountering
mainly for the 0x solver which is aggressively trying to match user
orders at the "arbitrage frontier", but then failing to settle those
solutions (as the liquidity is sniped by non CoW arb bots).

To mitigate this, we introduce the concept of a minimum surplus below
which we are not willing to bid on a given order. Similar to slippage
limits, we allow to configure the minimum surplus to be either relative
or absolute in terms of the order size.

## Changes
- [x] boilerplate logic to configure the new configuration parameters 
- [x] introduce the concept of min surplus as a strict copy of the
existing slippage types and logic (first commit)
- [x] refactor to combine min surplus and max slippage into a single
type (second commit)

## Test Plan

added unit tests