Remove deprecated callDataInput and queryBatchSwap (#159)
This PR removes deprecated fields from the Balancer SOR endpoint.
To give a bit more context, these fields were used to request the api
call to also perform an on-chain query to get up-to-date results.
This is no longer available, as the complexity of building the on-chain
call grew beyond the SOR responsibilities.
Balancer FE currently relies on the Balancer SDK to perform that
on-chain query with the swap path returned by the SOR endpoint.
This flow of performing an SOR request + on-chain call to get up-to-date
results is considered the ideal scenario to mitigate any issues with
stale/cached data the API might have.
I intend to create a separate PR adding this extra on-chain step to the
gnosis solver.

---------

Co-authored-by: ilya <ilya@cow.fi>