Balancer on-chain query swap (#161)
This goal of this PR is to add an intermediary on-chain query swap to
make sure amounts are up-to-date with on-chain state. Since SOR relies
on Balancer API data, which might be cached for a few seconds, amounts
may get slightly outdated. This shouldn't be an issue in most cases, but
it might become a problem for swap paths with wrap/unwrap steps and
slippage tolerance is too low (near zero).

The proposed solution adds an rpc_url as input to the Sor, so it's able
to perform the on-chain query swap internally. In case it's not able to
make the on-chain call by any reason, it falls back to using amounts
from the SOR quote response.

query_swap relies on a helper contract that is not yet present in your
infra, so I've added BalancerQueries to a [working branch on the
services
repo](https://github.com/brunoguerios/services/tree/add-balancer-queries-2).
Ideally we should be able to merge it there and update this current PR
to point to the official dependency once it's released. Please advise on
how to move forward with this.

Proper tests are still pending as I'm still trying to figure out the
best approach. Here are a few options:
1. Have some sort of integration tests that runs on a local node/fork
2. Mock the RPC calls
3. Update the implementation to rely on dependency injection to inject
something like a `query_swap_provider` to the Sor. With that we could
have an `OnChainQuerySwapProvider` for production code and a
`MockQuerySwapProvider` for tests. This would cause some extra side
effects in the code, but might be more elegant. Let me know if you'd
like for me to expand on that.

Please let me know if you have any suggestions on how to approach this.

PS: my Rust background is a bit limited, so feel free to challenge
anything you think could be improved - any feedback will be much
appreciated 🙏

---------

Co-authored-by: ilya <ilya@cow.fi>